+++ tests/e2e/manifest-lint/baseline.txt
@@ -35,7 +35,6 @@ create	tests/e2e/multinode/placement_nat_test.go	CreateVpc	#1
 create	tests/e2e/multinode/placement_nat_test.go	RunInstances	#1
 create	tests/e2e/multinode/placement_nat_test.go	RunInstances	#2
 create	tests/e2e/multinode/volume_durability_test.go	CreateVolume	#1
-create	tests/e2e/multinode/volume_test.go	CreateVolume	#1
 create	tests/e2e/multinode/vpc_test.go	CreateSubnet	#1
 create	tests/e2e/multinode/vpc_test.go	CreateVpc	#1
 create	tests/e2e/multinode/vpc_test.go	RunInstances	#1
