+++ tests/e2e/multinode/tests_test.go
@@ -9,19 +9,15 @@ import (
 // Top-level Test* wrappers, each delegating to a runX function. Parallel tests
 // share the read-only singleton trio; sequential tests mutate cluster state
 // (stop/start nodes, predastore, EIP pool) and are ordered so the cluster is
-// fully restabilised before GuestSSH probes every trio member.
+// fully restabilised before later tests probe the trio.
+//
+// Pre-flight (/dev/kvm writability, peer SSH reachability) runs inside
+// requireMultiNodeFixture itself, so the package fixture singleton fails
+// fast with a clear message before any Test* body executes.
 
-// TestMultinodePreflight runs sequentially first to initialise the package fixture singleton.
-func TestMultinodePreflight(t *testing.T) {
-	runPreflight(t, requireMultiNodeFixture(t))
-}
-
-// Baseline tests run before needInstanceTrio so the default SG/subnet/route table
-// are in pristine state. Both own dedicated SGs and self-cleaning instances.
-func TestMultinodeDefaultSGReachabilityBaseline(t *testing.T) {
-	runMultinodeDefaultSGReachabilityBaseline(t, requireMultiNodeFixture(t))
-}
-
+// Baseline test owns a dedicated SG and self-cleaning instances; it runs
+// before needInstanceTrio so the default SG/subnet/route table are in
+// pristine state.
 func TestMultinodeSameSGCrossHostComms(t *testing.T) {
 	runMultinodeSameSGCrossHostComms(t, requireMultiNodeFixture(t))
 }
@@ -29,11 +25,6 @@ func TestMultinodeSameSGCrossHostComms(t *testing.T) {
 func TestMultinodeClusterHealth(t *testing.T) {
 	t.Parallel()
 	runClusterHealth(t, requireMultiNodeFixture(t))
-}
-
-func TestMultinodeInstanceDistribution(t *testing.T) {
-	t.Parallel()
-	runInstanceDistribution(t, requireMultiNodeFixture(t))
 }
 
 // TestMultinodeDNS is sequential because it launches guests and briefly stops
@@ -49,11 +40,6 @@ func TestMultinodeJetStreamReplicas(t *testing.T) {
 	runJetStreamReplicas(t, requireMultiNodeFixture(t))
 }
 
-// TestMultinodeVolumeLifecycle is sequential: touches predastore state.
-func TestMultinodeVolumeLifecycle(t *testing.T) {
-	runVolumeLifecycle(t, requireMultiNodeFixture(t))
-}
-
 // TestMultinodeVolumeDurability is sequential and declared before CrossNodeOps
 // (which destabilises trio[0]'s sshd) so its guest-SSH probes hit settled VMs.
 // Touches predastore state.
@@ -67,8 +53,7 @@ func TestMultinodeCrossNodeGateway(t *testing.T) {
 	runCrossNodeGateway(t, requireMultiNodeFixture(t))
 }
 
-// TestMultinodeCrossNodeOps is sequential: stops/starts trio[0]; GuestSSH is declared
-// after this so it probes the trio only after the cluster has restabilised.
+// TestMultinodeCrossNodeOps is sequential: stops/starts trio[0].
 func TestMultinodeCrossNodeOps(t *testing.T) {
 	runCrossNodeOps(t, requireMultiNodeFixture(t))
 }
@@ -90,13 +75,6 @@ func TestMultinodeOVNRaft(t *testing.T) {
 // TestMultinodeSpread is sequential after NodeRecovery; owns 10.100.0.0/16 + EIP pool.
 func TestMultinodeSpread(t *testing.T) {
 	runSpread(t, requireMultiNodeFixture(t))
-}
-
-// TestMultinodeGuestSSH is sequential and declared last so it runs after the cluster has
-// fully restabilised. CrossNodeOps stops/starts trio[0] and only waits for "running" — not
-// sshd — so GuestSSH must not probe until the guest has settled.
-func TestMultinodeGuestSSH(t *testing.T) {
-	runGuestSSH(t, requireMultiNodeFixture(t))
 }
 
 // TestMultinodeVPCNetworking owns 10.200.0.0/16 (no EIP use); safe to run in parallel.
